% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterflow-class.R
\docType{class}
\name{AfterFlow-class}
\alias{AfterFlow-class}
\alias{exprs,AfterFlow-method}
\title{S4 AfterFlow (Virtual) Class}
\usage{
\S4method{exprs}{AfterFlow}(object)
}
\arguments{
\item{af}{AfterFlow object}
}
\value{
AfterFlow object
}
\description{
Virtual class that can be either an [AfterFlowFrame] or [AfterFlowSet],
both subclass [AfterFlowBase]

Resets AfterFlow to the original imported FCS file
}
\examples{
exprs(af) -> returns the MFI matrix of the active flow object
}
