% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterflow-class.R
\name{afterflow_check}
\alias{afterflow_check}
\title{AfterFlow Check}
\usage{
afterflow_check(af, invis = TRUE)
}
\arguments{
\item{af}{AfterFlow object}
}
\value{
Returns the primary class of the AfterFlow object
}
\description{
Makes sure the passed object is an AfterFlow object
}
\examples{
afterflow_check(af) -> If an AfterFlow object, nothing happens
}
