% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterflow-class.R
\name{afterflow_cluster}
\alias{afterflow_cluster}
\title{AfterFlowSet Clustering}
\usage{
afterflow_cluster(af, verbose = TRUE)
}
\arguments{
\item{af}{AfterFlow object}

\item{verbose}{Boolean for verbose message printing, default TRUE}
}
\value{
AfterFlow object
}
\description{
Clustering using FlowSOM
}
\examples{
afterflow_data(af) -> returns a current data
}
