% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterflow-class.R
\name{afterflow_data}
\alias{afterflow_data}
\title{AfterFlow Data}
\usage{
afterflow_data(af)
}
\arguments{
\item{af}{AfterFlow object}
}
\value{
AfterFlow object
}
\description{
Gets the active MFI values with the existing metadata
}
\examples{
afterflow_data(af) -> returns a current data
}
