% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterflow-class.R
\name{afterflow_dimred}
\alias{afterflow_dimred}
\title{AfterFlowSet Dimensional Reduction}
\usage{
afterflow_dimred(af, method = "umap", verbose = TRUE)
}
\arguments{
\item{af}{AfterFlow object}

\item{method}{Method for dimensional reduction, default umap}

\item{verbose}{Boolean for verbose message printing, default TRUE}
}
\value{
AfterFlow object
}
\description{
Clustering using FlowSOM
}
\examples{
afterflow_data(af) -> returns a current data
}
