% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterflow-class.R
\name{afterflow_filter}
\alias{afterflow_filter}
\title{AfterFlowSet Filter}
\usage{
afterflow_filter(af, meta, value, verbose = TRUE)
}
\arguments{
\item{af}{AfterFlow object}

\item{meta}{Metavariable}

\item{value}{Value to filter on}

\item{verbose}{Boolean for verbose message printing, default TRUE}
}
\value{
AfterFlow object
}
\description{
Filtering using FlowSOM
}
\examples{
afterflow_data(af) -> returns a current data
}
