% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterflow-class.R
\name{afterflow_markers}
\alias{afterflow_markers}
\title{AfterFlow Markers}
\usage{
afterflow_markers(af, filtered = FALSE, verbose = TRUE)
}
\arguments{
\item{af}{AfterFlowFrame object}

\item{verbose}{Boolean for verbose message printing, default TRUE}
}
\value{
AfterFlow object
}
\description{
Applies the compensation matrix to the AfterFlow object if available
}
\examples{
afterflow_markers(af) -> returns a compensated AfterFlow object
}
