% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterflow-class.R
\name{afterflow_reset}
\alias{afterflow_reset}
\title{AfterFlow Reset}
\usage{
afterflow_reset(af, verbose = TRUE)
}
\arguments{
\item{af}{AfterFlow object}

\item{verbose}{Boolean for verbose message printing, default TRUE}
}
\value{
AfterFlow object
}
\description{
Resets AfterFlow to the original imported FCS file
}
\examples{
afterflow_reset(af) -> returns a reset AfterFlow object
}
