% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterflow-class.R
\name{afterflow_scale}
\alias{afterflow_scale}
\title{AfterFlow Scale}
\usage{
afterflow_scale(af, verbose = TRUE)
}
\arguments{
\item{af}{AfterFlow object}

\item{verbose}{Boolean for verbose message printing, default TRUE}
}
\value{
AfterFlow object
}
\description{
Scaling data
}
\examples{
afterflow_data(af) -> returns a current data
}
