% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterflow-class.R
\name{afterflow_set_expr}
\alias{afterflow_set_expr}
\title{AfterFlow Set MFIs}
\usage{
afterflow_set_expr(af, expr, verbose = TRUE)
}
\arguments{
\item{af}{AfterFlow object}

\item{expr}{data.frame of MFI values, requires an "exp" variable if a Set}
}
\value{
AfterFlow object
}
\description{
Set MFI values manually
}
\examples{
afterflow_data(af) -> returns a current data
}
