% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterflow-class.R
\docType{class}
\name{AfterFlowBase-class}
\alias{AfterFlowBase-class}
\title{S4 AfterFlowBase (Super) Class}
\description{
This handles storage and manipulation of the FCS files and MFI values.
Subclassed by [AfterFlowFrame] and [AfterFlowBase]
}
\section{Slots}{

\describe{
\item{\code{fcs_orig}}{Original FCS file(s)}

\item{\code{fcs_active}}{Processed FCS file(s)}

\item{\code{metadata}}{tibble with metadata information, by event}

\item{\code{markers_exclude}}{List of excluded markers for analysis}

\item{\code{comp_key}}{Compensation/spillover matrix key (only set when applied)}

\item{\code{transform_method}}{Transformation method (empty character for none)}

\item{\code{batch_covar}}{Batch correction covariable}

\item{\code{batch_bool}}{Whether batch correction was applied to current active data}

\item{\code{clust}}{Vector of clustering methods run (corresponding to metadata columns)}

\item{\code{dimred}}{Vector of dimensional reduction methods run (corresponding to metadata columns)}

\item{\code{seed}}{Seed value}
}}

