% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterflow-class.R
\docType{class}
\name{AfterFlowFrame-class}
\alias{AfterFlowFrame-class}
\title{S4 AfterFlowFrame (Sub) Class}
\description{
Contains the original and processed [flowCore::flowFrame]
Subclass of [AfterFlowBase]
}
\section{Slots}{

\describe{
\item{\code{fcs_orig}}{Original flowFrame}

\item{\code{fcs_active}}{Processed flowFrame}
}}

