% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterflow-class.R
\name{afterflowframe}
\alias{afterflowframe}
\title{afterflowframe}
\usage{
afterflowframe(fcs_file, list_metadata, seed = 5841, verbose = TRUE)
}
\arguments{
\item{fcs_file}{Path of the FCS file to be processed}

\item{list_metadata}{Named vector of metadata variables (example: c(sample = "1"))}

\item{seed}{Seed to use for all random calculations, default 5841}

\item{verbose}{Whether to print status messages, default TRUE}
}
\value{
AfterFlow object
}
\description{
This handles storage and manipulation of the FCS files and MFI values.
}
\examples{
AfterFlow("fcs_location", c(batch = "1")) -> returns an AfterFlow object
}
