% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterflow-class.R
\docType{class}
\name{AfterFlowSet-class}
\alias{AfterFlowSet-class}
\title{S4 AfterFlowSet (Sub) Class}
\description{
Contains the original and processed [flowCore::flowSet]
Subclass of [AfterFlowBase]
}
\section{Slots}{

\describe{
\item{\code{fcs_orig}}{Original flowSet}

\item{\code{fcs_active}}{Processed flowSet}
}}

