% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterflow-class.R
\name{afterflowset}
\alias{afterflowset}
\title{afterflowset}
\usage{
afterflowset(
  af_fh,
  markers_exclude = c("AF-A", "dead"),
  seed = 5841,
  verbose = TRUE
)
}
\arguments{
\item{af_fh}{Path of the FCS file to be processed}

\item{markers_exclude}{Markers to exclude for analysis, defaults defined with
AF-A (autofluorescence) and dead}

\item{seed}{Seed to use for all random calculations, default 5841}

\item{verbose}{Whether to print status messages, default TRUE}
}
\value{
AfterFlow object
}
\description{
This handles storage and manipulation of the FCS files and MFI values.
}
\examples{
AfterFlow("fcs_location", c(batch = "1")) -> returns an AfterFlow object
}
