% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterflow-class.R
\name{afterflowset_batch}
\alias{afterflowset_batch}
\title{AfterFlowSet Batch Correct}
\usage{
afterflowset_batch(af, cv = NULL, verbose = TRUE)
}
\arguments{
\item{af}{AfterFlow object}
}
\value{
AfterFlow object
}
\description{
Gets the active MFI values with the existing metadata
}
\examples{
afterflow_data(af) -> returns a current data
  tib_fcs_corrected <- cyCombine::correct_data(
    tib_fcs,
    cyCombine::create_som(
      cyCombine::normalize(
        tib_fcs,
        markers = markers_include,
        norm_method = "scale"
      ),
      markers = markers_include,
      seed = seed
    ),
    covar = batch_covar,
    markers = markers_include
  )
}
