% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehandler-class.R
\docType{class}
\name{FileHandler-class}
\alias{FileHandler-class}
\title{S4 FileHandler Class}
\description{
This handles all directory and file operations, custom directories passed
during initialization
}
\section{Slots}{

\describe{
\item{\code{input}}{Input directory}

\item{\code{output}}{Output directory}

\item{\code{fcs}}{List of files, grouped by population, with vectors named by exp}

\item{\code{root}}{Root population}

\item{\code{metadata}}{Metadata for each file in data.frame form}
}}

