% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehandler-class.R
\name{filehandler_check}
\alias{filehandler_check}
\title{FileHandler Check}
\usage{
filehandler_check(fh)
}
\arguments{
\item{fh}{FileHandler object}
}
\value{
Nothing
}
\description{
Makes sure the passed object is a FileHandler object
}
\examples{
filehandler_check(fh) -> If an FileHandler object, nothing happens
}
