% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehandler-class.R
\name{filehandler_get_fcs}
\alias{filehandler_get_fcs}
\title{FCS Filenames}
\usage{
filehandler_get_fcs(fh, pop = character(), exp = character())
}
\arguments{
\item{fh}{FileHandler object}

\item{pop}{Optional population subset, default empty character}

\item{exp}{Optional experiment subset, default empty character}
}
\value{
List of vectors of filenames
}
\description{
Returns a list of vectors of FCS filenames split by populations
}
\examples{
filehandler_get_fcs(fh) -> Returns all fcs filenames by population
filehandler_get_fcs(fh, pop = "CD3+") -> Returns only CD3+ fcs filenames
filehandler_get_fcs(fh, exp = "Exp1") -> Returns only Exp1 fcs filenames across populations
}
