% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-functions.R
\name{show_message}
\alias{show_message}
\title{Show Message}
\usage{
show_message(msg, details = character(), final = "message")
}
\arguments{
\item{msg}{Message to print}

\item{details}{Optional details, if given, will print along with message}

\item{final}{Name of the end function to call, defaults to base message}

\item{lvls}{Integer to indicate number of indent spaces, default to 0}
}
\value{
Nothing
}
\description{
Prints a message that is passed with the timestamp.
Indents all lines beyond the first past the time stamp
Throws an error instead of just printing nicely if flag is set.
}
\examples{
show_message("This is a regular message") # Prints message with a timestamp
show_message("This is a warning message", final = warning) # Prints warning
show_message("This is an error message", final = stop) # Prints message with stop
}
